forge-get-repository: Change how forge.remote is used
If there is only one remote, then use that even if `forge.remote'
is a remote that does not exist.  If there are more remotes, then
use the one specified by `forge.remote' or else "origin", but if
`forge.remote' is set but the specified remote does not exist,
then no longer fall through to "origin".