Require subr-x
Magit used to do that, but now it only does so at compile-time.  We
could only require it at compile-time here too, but that would mean
that we had to do it in every file that uses some its macros and I
do not feel like doing that again right now.