forge--pull-notifications: Fix and refactor
The old implementation has bit-rotten and didn't work at all anymore.
First fixing that and then doing the refactoring in a second step
would be a waste of my time.

The new implementation only has to make two requests.  As before the
first request uses the REST v3 API because the GraphQL v4 API does not
expose notifications.  Then the response is transformed into a single
GraphQL request.  Previously we made a request for each notification.

Drop support for fetching the notifications for a single repository
because that would complicate the implementation and lead to outdated
data being displayed when listing the notifications.  It also is not
necessary anymore because fetching is much faster now.

Drop support for fetching all topics when encountering a notification
for a repository that the user previously pulled.  That's not useful
and would complicate matters a lot.

This also makes it possible to drop `forge--pull-repository's
SKIP-NOTIFICATIONS argument because we no longer have to avoid
infinite mutual recursion between that function and
`forge--pull-notifications'.