forge--update-{issue,pullreq}: Avoid foreign key constraint error
When storing a topic that was fetched because it was referenced in a
notification, then we cannot save the assignees and labels, because we
use the the assignee/label from the respective table of the repository
as a foreign key and that might not exist.

Eventually we should fetch the available repository's assigness and
labels along with the notifications, but that requires some major
changes.