forge--topic-buffer-lock-value: New magit-buffer-lock-functions fn
Without this the object would be used as lock-value.  After modifying
a topic, a new object is not `equal' to an outdated object for the
same topic, which would result in an existing buffer not being reused.

Split out some of the logic into a separate function, so that that can
be used separately.  Adjust some existing code to do so.