Take forge-alist into account when parsing urls
Currently we only support forges whose repositories are all available
at "HOST/OWNER/NAME[.git]", where neither OWNER nor NAME may contain
a slash.

In order to support semi-forges, which do not necessarily follow this
convention, we have to support urls that contain an arbitrary number
of slashes in the "path" part.

We already supported slashes in the "host" part before, in which case
"host" actually is a misnomer.  But now that slashes may also appear
in either in OWNER or NAME, not just as separators, this cannot be
expressed as a regular expression anymore.

Split the "OWNER/NAME" part using a generic function.  We do not use
a class slot such as `allow-slash-in' because not doing so gives as
additional flexibility, e.g. for one semi-forge we have also have to
remove "~" from the "owner" part.