forge--repository-ids: Drop support for name-based ids
Until recently it made (little) sense to offer purely name-based IDs
as an alternative style, because that made it possible to avoid an
API request.

However for Gitlab, which we recently started to support, we would
have to make such a request anyway because we really do need their
ID for that forge.

Drop the complication all together.