Treat Bitbucket as a semi-forge
Bitbucket actually is a proper forge whose API we do not support.

The one benefit that not-fully-supported forges currently have
over semi-forges is that forges expose pull-requests as refs in
the upstream repository.  Well, all forges except Bitbucket.

And it doesn't look good; this issue was opened six years ago:
https://bitbucket.org/site/master/issues/5814.