Support teahub.io host
Note that this host is in pre-alpha and that it is not clear whether
the operators are trustworthy.

I have only added this to `forge-alist' because I couldn't find any
other more trustworthy *and* persistent public instances but needed
something for testing purposes.  The Gitea maintainers plan to create
their own instance, see https://github.com/go-gitea/gitea/issues/1029.
Once that is available, I will probably remove teahub.io again.

I am also not happy about having to modify `forge--url-regexp' just
for this one host, which is another reason I might stop supporting it
when something better comes along.

Also remove the "gitea.local" host, which I used for testing early on.
Note that local installations of both Gitea and Gogs are available on
localhost:3000 by default.  So adding such a Gitea instance to the
default value of `forge-alist' made it a bit harder for people trying
out Gogs.