Make sure the `updated' slot is always bound
When pulling topics we avoid fetching topics again, that have not
changed since we last pulled, using `forge--topics-until'.  That
function returns the date of the least recently updated row from
the `issue' or `pullreq' table.

This assumes that we always fetch all updated or new topics, but
that is not the case.  `forge--pull-notifications' and the kludge
`forge--pull-pullreq' can leave caps.

Until now we dealt with that by not updating the `updated' column,
which leads to "slot `update' unbound" errors.

We will have to find a proper way to record until what date we have
pulled *all* issues and pull-requests, possibly using two new slots
on the repository object.

Until then we have to make sure that the `updated' slot of topics at
least is never unbound.  We do so by setting it to the date "0" if
it would otherwise be unbound.