forge--pull-notifications: Paginate GraphQL request
It is unknown how many topics we can fetch in one request, but so far
it seems like 100 is okay.  If we go over the magical threshold, which
may also vary depending on the phase of the moon, we still get an
uninformative error like:

  HTTP Error: 502, "Bad gateway", "/graphql",
  ((data . "null")
   (errors
    ((message . "Something went wrong while executing your query. \
  This may be the result of a timeout, or it could be a GitHub bug. \
  Please include `XYZ` when reporting this issue."))))

The correct response would probably be "504 Gateway Timeout".

Re #20.