Only offer open topics as completion candidates by default
Change all commands that previously offered all issues or all
pull-requests as completion candidates to only offer the open ones
by default.  In the future other commands might be added again that
do always offer all issues/pull-requests.

With a prefix argument offer all issues/pull-requests.  The topic
at point continues to be offered as the default choice even if it
is closed.

Closes #42.