Offer to reset existing branch when branching a pull-request
The old behavior to silently use "pr-N" instead of "named-branch" when
"named-branch" already existed was somewhat reasonable.  However, when
both of these branches already existed, then the error "Branch `pr-N'
already exists", was very confusing because one would have expected
"named-branch" to be used in the first place, not "pr-N".