forge--sql: New function
This new function does the same job as `forge-sql' did, but it takes
the database as a parameter.  This way, we can avoid re-doing
work (even though subsequent calls to `forge-db' are very fast).  This
doesn't have any effect with `forge-sql' itself, but the usefulness is
apparent with repeated queries.