forge--format-topic-id: Temporary performance kludge
Due to recent refactoring in #97 it is now too expensive to determine
whether a topic has to be prefixed with "!" instead of "#" (which is
the case if it is a Gitlab pull-request).

The problem is that we end up calling `(forge-get-repository t)' once
per pull-request.  Previously we did it once.

For the time being use "#" unconditionally.

Re #103.