Better kludge for recent performance regression
Due to recent refactoring in #97 it became too expensive to determine
whether a topic has to be prefixed with "!" instead of "#" (which is
the case if it is a Gitlab pull-request).

The problem is that we end up calling `(forge-get-repository t)' once
per pull-request.  Previously we did it once.

78fe8e744c1d8ecc847f86c2a04ead73d6612a29 "fixed" that by just always
using "#", which is not correct.

This commit restores using "!" when appropriate.  As it was done
originally the value is only calculated once and then passed around.
Because more functions are involved now that is a rather ugly fix.

The proper way to address this probably is the introduction of a
`forge-gitlab-pullreq' class and a `type-prefix' class slot, but the
former would mean that the database has to be updated.  I want to
think about it more before committing to this approach.