manual: Correct information about the bindings of the pull commands
Despite the names of these commands they are better thought of as
fetch commands than as pull commands.  Like Git fetch commands they
fetch data and store it somewhere (in this case in a database, not
using Git refs).  Pulling means updating HEAD, which these commands
obviously don't do.

That's why these commands have to be available from the fetch
transient and are only optionally available from the pull transient.