Add browse-* commands to *-list-mode
A better approach would be to make FORGE-BROWSE-DWIM work in list-mode
buffers or at least be able to generalize FORGE-LIST-BROWSE-* so that it
can be defined in the common-map, FORGE-TOPIC-LIST-MODE-MAP, but I'm
unsure of how to implement that.

FORGE-BROWSE-TOPIC doesn't work on list-modes due to
FORGE-TOPIC-AT-POINT not working on them.