forge-alist: Add entry for framagit.org
Framasoft is a well known non profit organisation (at least in France)
which offers a set of free (speech) software tools hosted for the well
being of public users. Details on https://framasoft.org/en/.

Framagit is one of the service they offer to the public which is a
common Gitlab instance. (Similar to Debian's Salsa instance).