forge--object-id: Ensure we have bytes before base64 encoding
Gitlab label names may contain unicode characters and for unrelated
reasons we are forced to use those names instead of the ids in this
case.  Until this commit that lead to "Multibyte character in data
for base64 encoding" errors.

Instead of adding the kludge here we could have spread it across
`forge--update-issue' and `forge--update-labels'.

Closes #139.

https://lists.gnu.org/archive/html/help-gnu-emacs/2012-04/msg00173.html
> base64 is an encoding which applies to *bytes*, not chars.  So you first
> need to encode your sequence of chars into a sequence of bytes using
> some coding-system.  I.e. apply `encode-coding-string/region'.
> And after that, you can use base64-encode-string/region.

See #160 for why we have to use the names of Gitlab labels in place
of their ids.

-- Jonas Bernoulli