forge--pullreq-ref: Never return a branch name
Instead always return a ref in the "refs/pullreqs/" namespace.  That
is what `forge--pullreq-ref-1' already did used.  `forge--pullreq-ref'
only had one caller and it should have used `forge--pullreq-ref-1' too.

Because it didn't we could ended up showing the log for
"master..master" if the contributor did not bother with a dedicated
feature branch.  But even if that it is not the case it is still
better to show "master..refs/pullreqs/N" instead of "master..NAME".

Change `forge--pullreq-ref' to behave like `forge--pullreq-ref-1'
and remove the latter.

Fixes #134.