forge-get-repository: Return the buffer-local repository
Change the method that previously returned a repository based on the
buffer-local value of `default-directory' to first try the values of
(the new) `forge-buffer-repository' or `forge-buffer-topic'.

This is necessary when a forge buffer is created from outside the
corresponding repository, then the inherited `default-directory' may
not have anything to do with the relevant forge (and Git) repository.

Another approach would be to ensure `default-directory' is "correct",
but that would require that we keep a mapping from forge repositories
to checkouts of the corresponding Git repositories.  That would not
work when no such checkout exists.

Closes #93.