forge--branch-pullreq: Abort on deleted Gitlab source branches
When the source branch of a merge-request on Gitlab is deleted,
then the respective merge-request ("refs/merge-requests/N") is
also deleted.

Theoretically we could preserve the corresponding ref locally,
but that would be very cumbersome so we don't.

It is even possible for this ref to be missing for open requests
because Gitlab does neither disallow deleting a branch if it is
associated request nor close that request automatically.