forge-checkout-pullreq: Just checkout existing branch
Previously if the branch already existed, then the user had to confirm
that they want to reset that branch.  Now the branch is simply checked
out as-is.

The benefit is that one can simply checkout out a pull-request branch
without knowing if it already exists.

The drawback is that if a branch by the expected name already exists
but wasn't actually created from the pull-request, then that won't be
detected and rectified.  However that should only happen rarely, so
this seems acceptable.

Closes #86.