forge--url-regexp, forge--branch-pullreq: Support plain http
This should make it possible to connect to insecure forges that are
only accessible over plain http.  The forge classes provided by this
package assume the use of https, so one has to copy such a forge
definition, rename it and replace "https" with "http" in order to
connect to an insecure instance.

Closes #9.