Use specified background color for labels
On Github and Gitlab at least, a label's color is used as its
*background* color.  Forge used this color as the foreground and box
color instead because of limitations of Emacs' display engine.  Magit
highlights the current section using an overlay that changes the
background color.  It is not possible to give an overlay (property)
a priority that is lower than that of a text property, so overlay
properties always override the corresponding text properties.

Back in the days tags and branches used to have a background color in
Magit but due to this issue we switched to using the box face property
instead, which we could do because the used color is under the control
of the package or theme author.

We did the same for labels, but because we don't have any control over
the used colors this could result in boxes that are barely visible.
So going forward we set the background color of labels and do so using
overlays with higher priority than the highlighting overlay.

Magit actually did the same for tags and branches for a while, but
stopped doing so because this is a rather crude hack.  But for labels
we do not have a better option, given that we have no control over the
used colors.