forge--branch-pullreq: Fix branch creation for existing remote
Hoist branch creation out of a conditional so that it happens both
when the remote already exists and when the remote needs to be
added.

Previously the branch was not created when the remote already existed
and this caused an error when this same function tried to configure
the branch.

Fixes #218.