forge-get-topic: Fix getting an issue using a number
Gitlab is a FORGE-2, it uses different namespaces for issues and
pull-request.  We are coping with that by using negative integers.

Because other forges don't do this `forge-get-topic' for natural
numbers continued to first try `forge-get-pullreq' and only if that
returned nil also try `forge-get-issue'.

But for Gitlab that is the wrong order so we flip it now.  We do the
same for other methods beside the integer method for consistency.

Closes #222.