Refresh tabulated-list-mode buffers when ready and only once
`tabulated-list-mode' and derived modes add `tabulated-list-revert' to
`display-line-numbers-mode-hook', which causes `tabulated-list-print'
to be called.

If `global-display-line-numbers-mode' is enabled, then enabling a
major-mode causes `display-line-numbers-mode-hook' to be called.

This means that merely enabling `tabulated-list-mode' causes the
buffer to be refreshed, which causes it to be printed.  We could work
around this by enabling the mode later, but that has the drawback that
the buffer is printed twice, which is wasteful.

So instead we prevent the mode from needlessly reverting.  This is
necessary because apparently `tabulated-list' assumes that the derived
mode sets all the relevant `tabulated-*' variables but we use the
various listing commands to do so instead because otherwise we would
have to define up to one mode per listing command just to conform to
this interface (aka accidental complexity).

Closes #229.