Promote forge-add-repository
Forge supports creating topics without first having to pull all
existing topics.  The only complication is that on has to use
`forge-add-repository' instead of `forge-pull' to initially add
the repository to the local repository.

But that command was not promoted enough leading to users requesting
this feature more than once: #230, #194, #121 and #102.  Now that the
command is prominently featured in the relevant places I hope we will
get fewer such requests.