Allow toggling Forge topics
Depending on the size of the notifications for the
repo, the redisplay of the magit-status buffer is
noticeably slower (in my particular case, the
magit-insert hooks for PRs and issues regularly
spend more than 2s each).

This PR allows the user to toggle redisplay for
these topics as needed, to avoid this performance
hit.

This commit implements the suggestions made by
@tarsius in the original PR.