manual: Use "upstream" as an alternative name for a remote
While I don't recommend using "upstream" as the name of the upstream
remote (one should just stick to "origin") that actually is the most
common alternative name, so it makes sense to show how to use that.

Avoid suggesting "main" as the name of a remote.  The name "main" has
been gaining traction as an alternative name for the branch commonly
known as "master".  Having both a branch and a remote named "main"
(even if it is in different repositories) would be quite unfortunate.