forge--branch-pullreq-{active,select}: New functions
Split up `forge--pullreq-branch' into two dedicated functions.

Previously there only was one function that returned a pull-request's
branch, which wasn't appropriate because it makes a huge difference
if we want to know what existing branch, if any, is associated with a
certain pull-request or what the best name would be when we are about
to create it.

Closes #292.