Correctly pull GitLab data with forge.omitExpensive
When `forge.omitExpensive` is set, some expensive fetches are avoided by
adding empty entries for them in a result variable.  However, this broke
handling the result of fetching the repository data: it was added a
level deeper than expected in the remaining code.

Fix this issue by looking at the result data: if the first element is a
cons cell it is the repository data and should be appended, otherwise it
is an additional fetch (issues, pull requests, ...) and should be pushed
to the results variable.