Fix handling of true forge.omitExpensive
Since [1: f5886b4b] the idea is to prevent some expensive requests
by injecting fake entries into the accumulated data that look like
we already fetched that data.

Due to how we use callback to chain requests we ended up injecting
those fakes to early, before handling the response of the initial
request and ended up not storing the data from that response.

Add the fake entry in the callback of the initial request, before
that calls the generic callback.

Fixes #347.

1: f5886b4b03689eb5929edef094a403d68b4d09d3
   forge.omitExpensive: New variable