forge-visit-topic: Fix regression
Since [1: ee558324] the interactive form passed an integer to the
function instead of an object, because the latter is a complex value
and that should be avoided if possible.

Unfortunately it is not possible here because this command is used
in contexts that are not confined to a single repository, so we have
to be explicitly about the repository.

Eventually we might have to change all commands to use objects instead
of integers again.

1: ee558324d3d494f614c46dab2f7701c32f740881
   forge-visit-topic: Maybe read topic in minibuffer