forge-forge.remote: Show the dynamic fallback value
Since [1: bc99603b] `forge.remote' no longer defaults to "origin".
Instead there is a dynamic fallback value, which depends on what
remote names are used in the current repository.

This accordingly adjusts the information that is being displayed
in the popup.

Depends on [2: 0bbf31a2ef] on the Magit repository.

Closes #482.

1: 2022-03-10 bc99603b5a0cd5b3a5f209772b7f818d205b8a3f
   forge-get-repository: Try harder to dwim and improve documentation

2: 2022-04-06 0bbf31a2efd769ac2a4734358b4da08a596e9514
   magit--git-variable:choices: Support fallback instead of default