Use existing pull-request branch as-is
If the local branch that represents a pull-request already exists,
then always use it as-is.  Never try to update it to point at the same
commit as the pull-request as it exists on a remote.  Trying to do the
latter would be a futile endeavor.  If the local and remote branches
have diverged, then the user is expected to reconcile that fact using
appropriate general purpose Magit command, which performs the action
that they feel is right in each particular case.  Since we configure
the fork remote and branch relationships as appropriate, it should be
easy enough to reason about diverging history.

We cannot know how to dwim if the user checks out / "creates" an
existing branch, which is/isn't the current branch in this/another
worktree, with history diverging on both sides / new commits only on
the local/remote side; without/with any staged and/or unstaged
changes; when we can/cannot use the same name as on the remote; the
fallback branch "pr-N" does/doesn't exist; the pull-request comes from
the upstream / a fork; pr metadata is/isn't already set up, and so on.

Closes #519.