forge--url-regexp: ignore "ssh." subdomains
Some instances of GitLab (and possibly other forges) serve their SSH
interface at a subdomain (e.g., "ssh.gitlab.some-company"). If the forge
is registered in forge-alist using the subdomain as GITHOST, then
forge-browse won't work because it derives URLs for the web interface
from the GITHOST. This commit causes forge to ignore the "ssh." prefix,
so that the forge can be registered with its web interface URL in
forge-alist (e.g., "gitlab.some-company").