Prevent post heading face from clobbering author & date faces
`add-face-text-property' results in this combination:

> There are text properties here:
>   face                 magit-diff-hunk-heading
>   font-lock-face       forge-post-author

This effectively hides `forge-post-author'; `forge-post-date' is
similarly affected.  With `font-lock-append-text-property', we get
instead:

> There are text properties here:
>   font-lock-face       (forge-post-author magit-diff-hunk-heading)

And this allows the author & date faces to appear.

Attributes from `magit-diff-hunk-heading-highlight' will still have
higher priority, as that face is applied via an overlay.  This means
that as the user moves from post to post, the author & date for the
post-at-point might e.g. lose their :foreground colors, if (a) they
define some (b) `magit-diff-hunk-heading-highlight' also defines some.

With Magit & forge's default values, this is not a problem.