Only pull "unseen" notifications
Ideally clients would be able to display the same notifications as are
by default listed at https://github.com/notifications, but because the
respective REST endpoint is bad and the respective GraphQL object does
not even exist, we cannot do that.

On the web, those notifications are listed that have not explicitly
been marked as "done" by the user.  The API only gives us the choice
to limit the list to notifications to those that the user has not even
looked at yet.  "done" != "unseen".

So previously we listed "all" last N notifications instead.  According
to the documentation that list ought to be sorted by "most recently
modified", which does not appear to be accurate.  As a result we got a
long list of seemingly random notifications.

Now we only pull and display "unseen" notifications.  That means that
if a user looks at a notification in a browser, then it won't later be
listed in Forge, unless the user explicitly marks the notification as
"unseen" after looking at in a browser.  Looking at a notification in
Forge does not mark it as "seen", so it remains in the list both when
viewed in Forge and a browser, until it is looked at in a browser.

This could be a useful and simple to implement feature, but because
of "limitations" of the API it is not.  When I implemented support
in Forge a few years ago, I hoped they would eventually fix the API.
They still haven't, and I have no intention to work around that on
the client side, if that is even possible without massive amounts of
polling.

Closes #578.