Separate read/this variants of visit commands
Change `forge-visit-{topic,issue,pullreq}' to always read the target,
even if on an appropriate topic.  Technically this is implemented in
Magit, by removing these commands from `magit-dwim-selection'.

Add new `forge-visit-this-{topic,repository}' commands.

Remove `forge-visit-repository' because now it would visit the current
repository, which doesn't make sense because the current repository is
the one that is already being visited.