Ignore thing-at-point when that would be misleading
Even in buffers that are not actually associated with any one
repository in particular, the value of `default-directory' may
still cause (forge-get-repository SYMBOL) to return a repository.

When there is a topic string at point, then we used the value that
that form returns, in various `forge-*-at-point' functions.  Now we
prevent that from happening, so that later sources are used instead,
which do return the correct repository, or a topic, from which the
correct repository can determined indirectly.