Store topic slugs in topic objects
A "topic slug" is the topic number prefixed with something like "#".

This would be trivial if all forges used the same namespace (i.e.,
consecutive integers) across all topic types.  Unfortunately Gitlab
decided to use a different namespace for pull-requests, which forces
them, and thus us, to use a different prefix, "!", for those.

Based on harmless prior art in [1: 43d5ca2a], this was further
abstracted in [2: 56a9f1ae] and mainly [3: 2019c224], using several
generic functions, which needlessly complicated matters a lot.

1: 2018-12-06 43d5ca2a031578f240258af28b058727d63cad50
   forge--topic-buffer-lock-value: New magit-buffer-lock-functions fn

2: 2019-02-18 56a9f1aee2d0f7cbf6fb45ea194a1db07e09c9d1
   forge--topic-type-prefix: Implement as generic function

3: 2019-02-18 2019c224aedd8bd2dbcf5935dae166b014e3ccae
   Generalize topic list insertion

This commit reverts that and instead stores the slug in the topic
object, and hardcodes "#" and "!" in a few places (namely functions
that get the topic at point based on the text at point).  Locating
these places can trivially be found, either by looking at this commit,
or by grepping, should the need to do work in this area ever arise
again.

Now that getting the slug has no additional cost compared to getting
the number, we don't have to determine it only once when listing
topics, and then pass it along as arguments between the involved
functions, to avoid a performance impact.  We already stopped doing
that in the parent of this commit.

This change also made it more appealing to prettify the completion
choices when reading a topic, because now that can be implemented
simply by removing some more code.

Also implement getting repository slugs ("OWNER/NAME") different.
Instead of manually building that string based on the two involved
slots in each place we need it, add a `slot-missing' method that
handles getting the value of the `slug' pseudo-slot.