forge--format(remote): Remove harmful method
Just use the `repository' method instead.  Translating from a
repository object to a remote string and then back to a repository
is unnecessary and dangerous because the second step may fail either
by returning no repository or the wrong repository, depending on
`default-directory'.

That caused the issue reported in #601.  To avoid similar bugs in
the future, remove the `remote' method altogether; if the repository
isn't known and cannot be determined, then the remote is meaningless.

Also remove `forge--split-remote-url' because it is trivial, and the
removed method was its only caller.