Track topic status only in issue and pullreq objects/table
Rename the slot `unread-p' to `status' for `forge-issue' and
`forge-pullreq'.  Move away from using a boolean because there
is a third value, which we will start using in the next commit.

Store the id instead of the number in the `topic' slot of
`forge-notification', and remove the `unread-p' and `forge' slots.

Since we have to change the schemata anyway, take the opportunity
to rearrange the `notification' table, and to add the `saved-p'
column proactively.