Use objects in topic listings
Due to prior experience with the performance impact of using objects
in Epkg's tabulated-lists, I chose to use raw rows instead.  The
impact is nowhere near as bad as I had feared, resp. how it used to
be, so given the advantages of using objects, I have changed my mind
about this.