Append topic labels to title in tabulated lists
We cannot use the same overlay-kludge here, which we use in Magit
buffers, so we have to avoid using the background color to highlight
the current column using `hl-line-mode', as that would cover the
background colors of the label.

Add a new face `forge-tablist-hl-line', which instead of using the
background color uses a box to mark the current line.

The function `forge--format-topic-labels' already existed but wasn't
used anywhere, so replace the implementation.  The old implementation
was apparently intended to avoid the conflict the other way around,
by using a box for the labels and letting the highlighting use the
background.  In a way that would be more consistent, but also look
worse.