forge-ls-recent-topics: Stop caching for now
We started caching in [1: 844aa6d5] without providing a reason, but
assuming it has something to with performance is reasonable.  This
cache has to be invalidated in some cases, but determining when that
is work for another day.  Until then we stop caching.  We might also
be able to speed things up, instead of caching.

In [2: 76596a2f] we added one case of cache invalidation.  Visiting
a topic should also invalidate the cache, so that status changes are
displayed in the list.  Instead of doing that, we disable the caching
altogether for now.

1: 2020-12-01 844aa6d52792dd56a14f1759312c96a3de9b988e
   Cache queries used by topic lists in status buffer

2: 2024-01-12 76596a2f0b55dfa409dd479d4205da3769d636b3
   forge-topic-list-order: Zap cache when set