forge--ghub-update-notification: Fix setting status
- Set the correct slot.
- Use `done', not `nil', as the "other" value.

`nil' used to mean "unknown but assume it is pending", but now we give
up on that complication.  We cannot know because the API sucks, and we
just have to live with having to pick one possible value.  As an added
benefit, `forge-topic-status-set-pending' no longer has to be more
complicated than the related commands.