Use "C-c C-m" binding for menu commands
Topics are also listed in Magit's status buffer and when point is
on a topic or topics section there, then the same key as elsewhere
should be used to invoke the respective menu command.  That cannot
be "L" because that would mean that `magit-log-refresh' would not
be available, just because point happens to be on a topic or topics
section.

I would have liked to use "C-c n", but that is not allowed because
"C-c <char>" is reserved for user bindings.  "C-c C-n" would have
been my next choice, but that is already used for commands that
create things ("n" for new).  So use "C-c C-m"; mnemonic "m" for
menu.