closql--db-update-schema: Fix updating from version 9 or older
We cannot use the object abstraction provided by closql while updating
table schemata because the functions that generate such objects assume
the latest schemata.  When updating from the second to last version,
then it can work (as it did when updating from 9 to 10), provided
columns are added first.  However, if a later version bump modifies a
relevant table, earlier update steps that use objects, are likely to
break.

Closes #625.
Closes #626.