forge-topic-mark-read: Only perform the appropriate status change
There are three statuses "unread", "pending" and "done".  The only
status change that makes sense when performing "mark as read" is to
change from "unread" to "pending".  "pending" and "done" topics are
already not-"unread", and the distinction between the two should not
be lost just because the user decides to look at the topic.

This function was accidentally not adjusted when we migrated to three
separate statuses, likely due to a rebase accident.