forge--split-url: Use same regexp for all forges
Previously we used a simpler regexp for Github, that only matched
a path part of the form "OWNER/REPO".  For other forges the path
part begins with "~" (which we have to strip) and/or the "OWNER"
part may have the form "OWNER/GROUP", in which case we use all of
that ("OWNER/GROUP") as the owner.

The regexp we previously used would have failed on invalid input such
as https://github.com/magit/forge/issues.  Now that error cannot be
detected at this stage anymore, causing a, potentially more confusing,
error later on.

However, due to where the value that is passed to this function comes
from, this isn't actually relevant.  It comes from Git configuration,
which too would be broken, if such an invalid value were used; and
that error would most likely have been noticed by the user earlier,
because it would make fetching and pushing impossible.