forge--format: Use webhost instead of githost
These values usually are the same, but when using ssh aliases, then
they are not.  Since this function is only used to format web URLs
(and other strings that are not URLs), the webhost has to be used.
The corresponding slot is named `forge', not `webhost', because it
is also used to identify the corresponding forge.