forge--read-topic: Allow lifting limit to active during completion
Previously the user could use a prefix argument to not limit
completion candidates to active topics.  Now they can instead lift
the while completion is already in progress.

This has the advantage that if one notices that the topic one wanted
to select unexpectedly is not among the completion candidates, one
can expand the candidates, instead of having to abort and invoke the
command again with a prefix argument.  It is also more discoverable,
because the key to expand the candidates is shown in the prompt.

Additionally, commands that use `forge-read-{topic,issue,pullreq}'
can now use the prefix argument for something else.