Only bind minibuffer-allow-text-properties in forge--read-topic
Previously globally consed `forge--topic-id' onto the existing value.
For some reason I believed this meant that only this property would be
allowed.  (For some other reason the global value was t for some user,
but I believe another package is responsible for that.)

Setting the value locally using `minibuffer-with-setup-hook' does not
appear to work (at least in Vertico), so use a let binding around the
call to `magit-completing-read' (which uses `completing-read').

Re #639.
Re https://github.com/minad/jinx/issues/140#issuecomment-1978951419.