forge-get-repository: Remove properties from string arguments
For the list method this is necessary to ensure propertized strings
are not stored in the database (because all values are stored in
their printed form).  This also takes care of other methods, which
ultimately call this method, and thus don't have to remove the
properties themselves.

For the :id method this is necessary to ensure the lookup is
successful.  This method cannot store a new repository in the
database.

Also see #639 and parent commit.