forge-{-read-topic,read-open-issue}: Do not rely on text properties
Don't rely on text properties to store the actual values inside the
strings the user selects from.  Using text properties is not reliable;
they might be removed by built-in completion.  Instead use an alist.
We have to replace that if the user extends the set of completion
candidates, but even then this is trivial.

Packages should not globally set `minibuffer-allow-text-properties'
as we did, because that can result in `completing-read' returning
propertized, which callers may not be prepared to handle.

Closes #639.