ghub--username(repo): Look up git variable in correct directory
The git variable may be set locally just for REPO.  If we are not
already in a working tree of REPO, then try to change to the one
recorded in its `worktree' slot (i.e., the one last visited), to
increase the odds that we get the correct value.

Re #646.