forge--topics-until: Drop UNTIL argument
The UNTIL argument was added in [1: 5efc0130].  If the repository
is sparse (i.e., hasn't been fetched before), then that is returned,
whether it is nil or not.  If it isn't nil, then only topics since
that date are fetched.

The UNTIL date (which is forwarded here by `forge--pull') must be
ignored, when fetching a repository again, to avoid gaps in the list
of topics (as opposed to a single stretch of the oldest topics being
ignored).

Doing this isn't strictly necessary as it would constituted a bug in a
caller of `forge--pull'.  Handle it accordingly, by placing an assert
in that function.

1: 2019-01-20 5efc013024bacf18367b45af6b8cba38cba24162
   Allow limiting initial pull to recently updated topics