forge-repository: Replace sparse-p slot with condition slot
There are three conditions a repository can be in `tracked', `known'
(in the database but only very limited data is being tracked), and
`stub' (not in the database and `id' is inaccurate).  A boolean
doesn't quite cut it.

The similarity of their names caused me to confuse the `sparse-p'
and `selective-p' slots more than once.  Preventing that is another
advantage of this change.