Implement new status update heuristics
No longer set the topic status when updating the corresponding
notification but when updating the topic itself.

Both parts of the API endpoints are defective in the same way; the
express the tristate unread/pending/done using a boolean, forcing
us to use heuristics.

Unfortunately, the new heuristics in `forge--update-status' make
it necessary for callers have to pass in additional information.

The new heuristics may need further tweaking, but to figure that
out we have to observe the new behavior for a while.

(When updating the notification, set the notifications `updated'
slot to the value of the `updated_at' field again.  If the value
is incorrect we will needlessly fetch the notification again, but
that should be the old effect of the potentially incorrect API
data.)